+++ src/lib/quickPlace.ts
@@ -87,10 +87,13 @@ function manualBoxAsPlacedBox(box: ManualPlacedBox): PlacedBox {
     canRotate: box.canRotate ?? true,
     stackable: box.stackable ?? true,
     maxStackLayers: box.maxStackLayers,
+    groundOnly: box.groundOnly,
     physicalLayer: 1,
+    verticalLayer: 1,
     workStep: 1,
     supportType: 'floor',
     supportedBy: [],
+    verticalSupportedBy: [],
   }
 }